+++ libs/angular-components/CHANGELOG.md
@@ -1,3 +1,43 @@
+# [@tehik-ee/tedi-angular-v2.2.0-rc.10](https://github.com/TEHIK-EE/tedi-design-system/compare/angular-2.2.0-rc.9...angular-2.2.0-rc.10) (2025-07-09)
+
+
+### Features
+
+* **community-multiselect:** Translate select-all and add no-options case [#640](https://github.com/TEHIK-EE/tedi-design-system/issues/640) ([ef8064d](https://github.com/TEHIK-EE/tedi-design-system/commit/ef8064d6dcca8b9785cb29e9be2024e50f9ed91e))
+
+# [@tehik-ee/tedi-angular-v2.2.0-rc.9](https://github.com/TEHIK-EE/tedi-design-system/compare/angular-2.2.0-rc.8...angular-2.2.0-rc.9) (2025-07-09)
+
+
+### Features
+
+* **community-multiselect:** Fix rc8 not building [#640](https://github.com/TEHIK-EE/tedi-design-system/issues/640) ([d8c7cc2](https://github.com/TEHIK-EE/tedi-design-system/commit/d8c7cc279b2062cc273e9699e06245304b68dabd))
+
+# [@tehik-ee/tedi-angular-v2.2.0-rc.8](https://github.com/TEHIK-EE/tedi-design-system/compare/angular-2.2.0-rc.7...angular-2.2.0-rc.8) (2025-07-08)
+
+
+### Features
+
+* **community-multiselect:** Add clearable tags method to clear options [#640](https://github.com/TEHIK-EE/tedi-design-system/issues/640) ([e731f8b](https://github.com/TEHIK-EE/tedi-design-system/commit/e731f8bef88673e33dd1418f128a3a8ec696b2c0))
+* **community-multiselect:** Added a comment [#640](https://github.com/TEHIK-EE/tedi-design-system/issues/640) ([dc2c5ce](https://github.com/TEHIK-EE/tedi-design-system/commit/dc2c5ce09e48f3cee0d136401ad114cd8a99604f))
+* **community-multiselect:** Code review fixes [#640](https://github.com/TEHIK-EE/tedi-design-system/issues/640) ([0f64173](https://github.com/TEHIK-EE/tedi-design-system/commit/0f64173c8a9b9b28d71bb81f9b2688b328c49d3e))
+* **community-multiselect:** Fix "select-all" and "select-group" keyboard navigation [#640](https://github.com/TEHIK-EE/tedi-design-system/issues/640) ([1a43ec5](https://github.com/TEHIK-EE/tedi-design-system/commit/1a43ec5ec254b4432b86725316327dd9a7848ccc))
+* **community-multiselect:** rewrite select to use cdkListbox and cdkOverlay [#640](https://github.com/TEHIK-EE/tedi-design-system/issues/640) ([61857c9](https://github.com/TEHIK-EE/tedi-design-system/commit/61857c90df0fb387b17bbf692114ca5cd637b040))
+
+# [@tehik-ee/tedi-angular-v2.2.0-rc.7](https://github.com/TEHIK-EE/tedi-design-system/compare/angular-2.2.0-rc.6...angular-2.2.0-rc.7) (2025-07-08)
+
+
+### Bug Fixes
+
+* **angular-animations:** move to peer dependency, change get started story [#911](https://github.com/TEHIK-EE/tedi-design-system/issues/911) ([377c8e1](https://github.com/TEHIK-EE/tedi-design-system/commit/377c8e198b2a8dd05c68007f6d6b6881816ecd3d))
+
+# [@tehik-ee/tedi-angular-v2.2.0-rc.6](https://github.com/TEHIK-EE/tedi-design-system/compare/angular-2.2.0-rc.5...angular-2.2.0-rc.6) (2025-07-07)
+
+
+### Bug Fixes
+
+* **modal:** Better content projection for modal header [#776](https://github.com/TEHIK-EE/tedi-design-system/issues/776) ([2c07120](https://github.com/TEHIK-EE/tedi-design-system/commit/2c071204887aabff93a35975765b11e1e7a916d5))
+* **modal:** Fixed modal content projection [#776](https://github.com/TEHIK-EE/tedi-design-system/issues/776) ([de0f77a](https://github.com/TEHIK-EE/tedi-design-system/commit/de0f77a7f34cb62194f36ba65d4aec281b791999))
+
 # [@tehik-ee/tedi-angular-v2.2.0-rc.5](https://github.com/TEHIK-EE/tedi-design-system/compare/angular-2.2.0-rc.4...angular-2.2.0-rc.5) (2025-07-04)